refactor: separating functionality into different components
